Updated ShutdownTask

+++ src/main/java/torimia/client/crosswordclient/ShutdownTask.java
@@ -3,6 +3,7 @@ package torimia.client.crosswordclient;
 import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
+import org.bson.Document;
 import torimia.client.crosswordclient.version1.service.MongoService;
 
 @Slf4j
@@ -16,8 +17,11 @@ public class ShutdownTask extends Thread {
     @Override
     public void run() {
         Thread.sleep(1000);
-        if (mongoService.deleteGame(gameId) == 1) {
-            log.info("Game with id \"{}\" deleted", gameId);
-        } else log.info("Game with id \"{}\" not deleted", gameId);
+        Document game = mongoService.findById(gameId);
+        if ((game != null) && (!game.get("status").equals("FINISHED"))) {
+            if (mongoService.deleteGame(gameId) == 1) {
+                log.info("Game with id \"{}\" deleted", gameId);
+            } else log.info("Game with id \"{}\" not deleted", gameId);
+        }
     }
 }
